shared: Disable hazard pointer destructor for now
The assertion that all HP records have been freed can fail when the
thread's state has not been closed, for example when the thread that was
responsible for closing that state has crashed.  Disable the destructor
for now, it's not important.  An alternative cleanup mechanism may be
implemented in the future to avoid unclosed states surviving.