Add pfetch sample
This sample demonstrates a few useful concepts in action.  The script
spins up a worker thread pool reading tasks from a SPMC ring FIFO and
writing status updates to a MPSC ring FIFO.  The enqueue and dequeue
operations are lock-free (concurrent) and use event counters for
efficient blocking when empty/full.  Progress is displayed with a
bsddialog interface.

For now the script is pointed at a directory containing a few large
files on my ZFS pool and copies the files to /dev/null as a mock for
work.  Eventually some portable mock setup would be nice.